Refactoring to separate controller and route for collection
