+++ src/backend/controllers/api/user.js
@@ -13,9 +13,3 @@ export const getUserCollections = (request, reply) =>
 
 export const getUserWishLists = (request, reply) =>
   getServerMethod('getUserWishListsByUserId')(request)(getIdRequestParam(request), callback(reply));
-
-export const getGamesInCollection = (request, reply) =>
-  getServerMethod('getGamesByCollectionId')(request)(getIdRequestParam(request), callback(reply));
-
-export const getGamesInWishlist = (request, reply) =>
-  getServerMethod('getGamesByWishlistId')(request)(getIdRequestParam(request), callback(reply));
